// src/@types/last_fm_artist_info_api.d.ts
// Generated by https://quicktype.io

import { LastFMError } from './last_fm_api';

export type LastFmArtistInfoAPI =
  | {
      artist: LastFmArtistInfoAPIArtist;
    }
  | LastFMError;

export interface LastFmArtistInfoAPIArtist {
  name: string;
  mbid: string;
  url: string;
  image: Image[];
  streamable: string;
  ontour: string;
  stats: Stats;
  similar: Similar;
  tags: Tags;
  bio: Bio;
}

interface Bio {
  links: Links;
  published: string;
  summary: string;
  content: string;
}

interface Links {
  link: Link;
}

interface Link {
  '#text': string;
  rel: string;
  href: string;
}

interface Image {
  '#text': string;
  size: string;
}

interface Similar {
  artist: ArtistElement[];
}

interface ArtistElement {
  name: string;
  url: string;
  image: Image[];
}

interface Stats {
  listeners: string;
  playcount: string;
}

interface Tags {
  tag: Tag[];
}

interface Tag {
  name: string;
  url: string;
}

export type SimilarArtist = {
  name: string;
  url: string;
  artistData?: Artist;
};